Update docs and structure for 1.4rc
