+++ AGENTS.md
@@ -1,6 +1,13 @@
+# Copernican Suite Development Guide (v1.4rc)
+
+This document supersedes `doc.json` and consolidates all development protocols, model interface definitions, and plotting guidelines.
+
+## Full Technical Specification
+
+```json
 {
   "projectName": "Copernican Suite",
-  "projectVersion": "1.3",
+  "projectVersion": "1.4rc",
   "lastUpdated": "2025-06-11",
   "description": "This document serves as the master technical specification and AI development interface for the Copernican Suite. It defines the required structure for all components, outlines the development protocol, and provides a 'how-to' guide for creating new cosmological model plugins. This version includes AI-hardening features to ensure strict adherence to the plugin interface.",
   "projectSchema": {
@@ -11,7 +18,7 @@
     "lcdm_model.py": "The reference implementation of the standard LambdaCDM model. Serves as the primary 'control' model and a template for plugins requiring numerical integration.",
     "usmf3b.py": "An example of a fully analytic cosmological model. Serves as a template for plugins with computationally simple, closed-form solutions.",
     "README.md": "The main user-facing documentation, providing a high-level overview, history, and future vision.",
-    "doc.json": "This file. The authoritative technical specification for AI and developer reference.",
+    "AGENTS.md": "This file. The authoritative technical specification for AI and developer reference.",
     "output/": "The dedicated directory where all generated plots, logs, and CSV files are saved."
   },
   "modelPluginInterface": {
@@ -80,13 +87,13 @@
       "rules": [
         "Add a 'DEV NOTE' at the top of any modified file summarizing changes for the current version and providing guidance for future modifications.",
         "Comment code extensively. Explain the purpose of new functions, the logic behind complex algorithms or bug fixes, and the flow of data.",
-        "Update the main README.md and this doc.json file to reflect the latest changes, architectural decisions, and future plans."
+        "Update the main README.md and AGENTS.md files to reflect the latest changes, architectural decisions, and future plans."
       ]
     },
     "creatingNewModelWorkflow": {
       "description": "The workflow for an AI to generate a new, compatible cosmological model plugin from a user's idea.",
       "steps": [
-        "Step 1: Ingest User Idea & Templates. Receive the user's conceptual model idea, this doc.json file, lcdm_model.py (numerical template), and usmf3b.py (analytic template).",
+        "Step 1: Ingest User Idea & Templates. Receive the user's conceptual model idea, this AGENTS.md file, lcdm_model.py (numerical template), and usmf3b.py (analytic template).",
         "Step 2: Formulate Equations. Translate the user's idea into a set of mathematical equations. Prioritize creating analytic, closed-form solutions for distance measures to ensure high computational performance. If numerical integration is unavoidable, follow the pattern in lcdm_model.py.",
         "Step 3: Create the Markdown (.md) Definition File. Following the `modelPluginInterface.markdownDefinitionFile` specification, write the complete .md file. This file is the 'source of truth'.",
         "Step 4: Create the Python (.py) Implementation File. Following the `modelPluginInterface.pythonImplementationFile` specification, write the Python code. Implement all mandatory functions and metadata variables using the exact names specified.",
@@ -108,3 +115,15 @@
     }
   }
 }
+
+```
+
+## AI Development Laws
+
+To any AI modifying this project:
+
+1. **Add a `DEV NOTE` at the top of each changed file** summarizing the modifications.
+2. **Comment code extensively** to explain the reasoning behind complex logic.
+3. **Update all documentation files**, including this `AGENTS.md` and `README.md`, whenever the codebase changes.
+
+Failure to follow these guidelines jeopardizes the maintainability of the Copernican Suite.
