+++ data/sne/tablef3.dat
