+++ models/__init__.py
@@ -1,0 +1,1 @@
+# DEV NOTE (v1.4rc): Package init for model plugins.
